+++ project/CRUD/discussions.php
@@ -5,7 +5,7 @@ $id = $title = $posted_by = $posted_on = $content = $group_id = $group_name = $c
 $user_id = $_SESSION['user_id'];
 
 // ADD
-if (isset($_POST['discussion_add'])) {
+if (isset($_POST['add_discussion'])) {
 
     // receive all input values from the form
     $title = mysqli_real_escape_string($conn, $_POST['title']);
@@ -26,7 +26,7 @@ if (isset($_POST['discussion_add'])) {
 
     if (count($errors) == 0) {
         $add = "INSERT INTO discussion (title, content, posted_by_uid, posted_on, group_id)
-            VALUES('$title', '$content', '$user_id', CURRENT_TIMESTAMP,'$group_id')";
+            VALUES('$title', '$content', '$user_id', NOW(),'$group_id')";
 
         if (mysqli_query($conn, $add)) {
             array_push($success, "Added successfully");
@@ -104,7 +104,7 @@ if (isset($_GET['delete_id'])) {
     $discussions = mysqli_query($conn, $query);
 
     ?>
-    <h3>Discussions</h3>
+    <h2>Discussions</h2>
     <hr>
     <table>
         <thead>
@@ -116,7 +116,7 @@ if (isset($_GET['delete_id'])) {
                 <th>Posted on</th>
                 <th>Group Name</th>
                 <th>Course Name</th>
-                <?php !isStudent() ? print '<th colspan="2">Action</th>' : ''; ?>
+                <?php isAdmin() ? print '<th>Action</th>' : ''; ?>
 
             </tr>
         </thead>
@@ -128,7 +128,7 @@ if (isset($_GET['delete_id'])) {
                 $title = $row['title'];
                 $content = $row['content'];
                 $posted_by = $row['username'];
-                $posted_on = $row['posted_on'];
+                $posted_on = date_convert($row['posted_on']);
                 $group_id = $row['group_id'];
                 $group_name = $row['group_name'];
                 $course_name = $row['course_name'];
@@ -137,15 +137,14 @@ if (isset($_GET['delete_id'])) {
                     <?php if (isAdmin()) {
                         echo '<td>' . $id . '</td>';
                     } ?>
-                    <td><?php echo $title ?></td>
-                    <td><?php echo $content ?></td>
-                    <td><?php echo $posted_by ?></td>
-                    <td><?php echo $posted_on ?></td>
-                    <td><?php echo $group_name ?></td>
-                    <td><?php echo $course_name ?></td>
-                    <?php if (!isStudent()) {
-                        echo '<td><a href="?page=discussions&update_view=true&update_id=' . $id . '">Update</a></td>';
-                        echo '<td><a href="?page=discussions&delete_view=true&delete_id=' . $id . '">Delete</a></td>';
+                    <td><?= $title ?></td>
+                    <td><?= $content ?></td>
+                    <td><?= $posted_by ?></td>
+                    <td><?= $posted_on ?></td>
+                    <td><?= $group_name ?></td>
+                    <td><?= $course_name ?></td>
+                    <?php if (isAdmin()) {
+                        echo "<td><a href='?page=discussions&delete_view=true&delete_id=" . $id . "' onclick='return confirm(&quot;Are you sure you want to delete?&quot;)'>Delete</a></td>";
                     } ?>
                 </tr>
             <?php
@@ -153,165 +152,5 @@ if (isset($_GET['delete_id'])) {
             ?>
         </tbody>
     </table>
-
-    <?php if (!isStudent()) { ?>
-        <a href="?page=discussions&add_view=true">
-            <button>Add Discussion</button>
-        </a>
-
-        <?php if (isset($_GET['add_view'])) { ?>
-            <hr>
-            <div class="form-container">
-                <form class="form-body" action="" method="POST">
-
-                    <?php
-                    echo display_success();
-                    echo display_error();
-                    ?>
-
-                    <h4><u>Add Discussion</u></h4>
-
-                    <div class="form-input">
-                        <label>Title</label>
-                        <span><input type="text" name="title"></span>
-                    </div>
-
-                    <div class="form-input">
-                        <label>Content </label>
-                        <br>
-                        <textarea name="content"></textarea>
-                    </div>
-
-                    <!-- <div class="form-input">
-                        <label for="course_id">For Course</label>
-                        <span>
-                            <select name="course_id" onchange=showData(this)>
-
-                                <option value="" selected hidden>Choose a Course</option>
-                                <?php
-                                // $courses = get_table_array('course');
-                                // foreach ($courses as $row) {
-                                //     $course_id = $row['course_id'];
-                                //     $course_name = $row['course_name'];
-                                //     echo "<option name=course_id id=course_id value='$course_id'>$course_name</option>";
-                                // }
-                                ?>
-                            </select>
-                        </span>
-                    </div> -->
-
-                    <div class=" form-input">
-                        <label for="group_id">For Group</label>
-                        <span>
-                            <select name="group_id">
-                                <option value="" selected hidden>Choose a Group</option>
-                                <?php
-
-                                $query = "SELECT * FROM student_group as g
-                                            JOIN group_of_course as gc ON gc.group_id = g.group_id
-                                            JOIN course as c ON c.course_id = gc.course_id";
-                                $groups = mysqli_query($conn, $query);
-
-                                foreach ($groups as $row) {
-                                    $group_id = $row['group_id'];
-                                    $group_name = $row['group_name'];
-                                    $course_name = $row['course_name'];
-                                    echo "<option name=group_id id=group_id value='$group_id'>$group_name - $course_name</option>";
-                                }
-                                ?>
-                            </select>
-                        </span>
-                    </div>
-
-                    <div class="form-submit">
-                        <input type="submit" name="discussion_add" value="Add">
-                    </div>
-
-                </form>
-            </div>
-
-        <?php } ?>
-
-        <?php if (isset($_GET['update_view'])) { ?>
-
-            <?php
-            $id = mysqli_real_escape_string($conn, $_GET['update_id']);
-            $query = "SELECT d.*, u.username, g.group_name, c.course_name FROM discussion as d
-                        JOIN users as u ON d.posted_by_uid = u.user_id
-                        JOIN student_group as g ON g.group_id = d.group_id
-                        JOIN group_of_course as gc ON gc.group_id = g.group_id
-                        JOIN course as c ON c.course_id = gc.course_id
-                        WHERE d.discussion_id='$id'
-                        ORDER BY discussion_id ASC";
-            $results = mysqli_query($conn, $query);
-
-            while ($row = mysqli_fetch_assoc($results)) {
-                $id = $row['discussion_id'];
-                $title = $row['title'];
-                $content = $row['content'];
-                $group_name = $row['group_name'];
-                $course_name = $row['course_name'];
-                // $update_group_id = $row['group_id'];
-            }
-            ?>
-
-            <hr>
-            <div class="form-container">
-                <form class="form-body" action="" method="POST">
-
-                    <?php
-                    echo display_success();
-                    echo display_error();
-                    ?>
-
-                    <h4><u>Update Discussion</u></h4>
-
-                    <div class="form-input">
-                        <label>Group Name</label>
-                        <span><?= $group_name ?></span>
-                        <label>Course Name</label>
-                        <span><?= $course_name ?></span>
-                    </div>
-
-                    <div class="form-input">
-                        <label>Title</label>
-                        <span><input type="text" name="title" value='<?= $title ?>'></span>
-                    </div>
-
-                    <div class="form-input">
-                        <label>Content</label>
-                        <br>
-                        <textarea name="content"><?= $content ?></textarea>
-                    </div>
-
-                    <!-- <div class="form-input">
-                        <label for="course">For Group</label>
-                        <span>
-                            <select name="group_id">
-                                <?php
-                                // $groups = get_table_array('student_group');
-                                // foreach ($groups as $row) {
-                                //     $group_id = $row['group_id'];
-                                //     $group_name = $row['group_name'];
-                                //     if ($update_group_id == $group_id) {
-                                //         echo "<option name=group_id value='$group_id' selected>$group_name</option>";
-                                //     } else {
-                                //         echo "<option name=group_id value='$group_id'>$group_name</option>";
-                                //     }
-                                // }
-                                ?>
-                            </select>
-                        </span>
-                    </div> -->
-
-                    <div class="form-submit">
-                        <input type="submit" name="update_discussion" value="Update">
-                    </div>
-                </form>
-            </div>
-
-        <?php } ?>
-
-    <?php } ?>
 
 </div>
